Update LICENSE, README and others to meet standard (#64)
This updates the LICENSE and README to meet [Defra's standard](https://github.com/DEFRA/dst-guides/blob/master/github/new_projects.md#add-a-licence-and-readme). It also updates the README to match how the app currently works.

As part of this we have added a `.env.example` file as per convention. Plus we have updated the .gitignore to be more comprehensive and consistent with other projects.