Add node nvm version file to project (#66)
As developers needing to maintain multiple services using different versions of node we need to be able to switch between. We also need some record of what version of node is actually being used by the app.

A popular tool to manage multiple Node versions is [nvm](https://github.com/nvm-sh/nvm). It has a feature where if a file named `.nvmrc` is present in a folder calling `nvm use` will automatically switch you to that version.

The `.nvmrc` file in the project also serves as an indicator of the version we are currently using.