Upgrade solution to use Node v12 (#101)
https://trello.com/c/bMR5DtRv

It was flagged in recent PEN testing that when the Docker image is built some libraries and packages still remain out of date, even with an `apk update && apk upgrade` command in the Dockerfile.

We believe it might be linked to the fact we are using an image based on Node v10. We already know support for v10 will stop in April 2021 so it's a good idea to the project onto using v12 which is the current LTS version of Node.

These changes cover everything need to move the project to use Node v12.