Delete 'unused' code (#103)
This project has _a lot_ of unused code. The majority of it is commented out. Some are files that duplicate something else but have `*-old*` somewhere in the title.

This change is all about removing this stuff. It has no value to the new team and just gets in the way.