Run npm audit fix against the project (#107)
Prior to this change running `npm install` would display the following message

```bash
added 23 packages, removed 17 packages, updated 56 packages and audited 845 packages in 11.732s

38 packages are looking for funding
  run `npm fund` for details

found 55 low severity vulnerabilities
  run `npm audit fix` to fix them, or `npm audit` for details
```

This change is a result of running `npm audit fix.

For reference, after doing so we're left with just one low severity vulnerability which is related to Airbrake. We already know we are using an deprecated version so will resolve the vulnerability when we update airbrake.