Remove unused transaction queue controller (#110)
Spotted whilst doing general housekeeping on the code (see PR's #101, #103, #104, #105, #107, #108) that we have the `app/controllers/v1/transaction_queue_controller.js` but never actually require it in the router.

We know it's not being used by WRLS so this change deletes it (along with any related code only referenced in the controller) to reduce the maintenance burden of the project.