Remove endpoint suppression (#109)
https://trello.com/c/2A5QEUMI

Back in [PR #77](https://github.com/DEFRA/charging-module-api/pull/77) and [#69](https://github.com/DEFRA/charging-module-api/pull/69) we added some functionality to tag routes, and then filter which actually got added using config taken from environment variables.

The context was we were about to go into PEN testing and a number of endpoints were still a work in progress. That work in progress was SROC related and as we have now taken the decision to focus that work in [sroc-charging-module-api](https://github.com/defra/sroc-charging-module-api) there is no longer a need to suppress it. This is especially so as we have been busy deleting it in PR's #104 and #108 !

Essentially this reverts things back to how they were before.