Add disinfect plugin to santize inputs to the API (#119)
https://trello.com/c/cjqqnLGs

During PEN testing of the service, it was found that it is susceptible to XSS. This was because inputs to the API were not properly being sanitized. Rather than roll our own sanitizer, we're opting to use a plugin on this project.

[Disinfect](https://github.com/genediazjr/disinfect) has been used on a number of Defra projects. It applies Google's [Caja](https://github.com/google/caja) HTML Sanitizer on route query, payload, and params.

This change adds it to this project.

N.B. For reference, if an input has an issue Disinfect will remove the problem value. It is on whatever receives it to decide whether the value being missing is cause to throw an error or not. Disinfect doesn't throw an error.