Create an `/admin` root path (#121)
We only really want to worry about the issues of versioning when it comes to public endpoints. Currently, we have some endpoints which are purely for the development team mixed up under `/v1` along with endpoints for client use.

To make things clearer we're moving those that are there purely for administrating the API under a root `/admin` path.

This includes the tests. Ideally, the structure of the test folder should be consistent with the structure of the actual folder. It just makes it easier to tie things up.