Move Airbrake controller to health path (#122)
Along with the move of the admin endpoints to an `/admin` path we're now focusing on those related to the health of our environment.

Currently, the Airbrake controller, which is used to confirm that error logging is working in an environment, is on the `/status` path.

We now appreciate there is a mixing of convictions; you need to be an admin to access `/status/airbrake` but anyone can access `/status`. It would also be useful to have similar endpoints that confirm our connection to the database and to the rules service.

So to make things clearer and resolve the mixing of convictions this change creates a `/health` root path and moves Airbrake to it. We can then add those other health checks in later changes.