Move /health under /admin (#124)
One final tweak to the path design. Following a principle of making it clear which routes need admin access, this change moves the new `/health` path under `/admin`. That way we now hopefully have a consistent principle. If it needs admin access it lives in the `/admin` path though we use sub-paths to categorise endpoints.