Fix issue with blown integers (#120)
https://trello.com/c/RIATlpAS

We had a report of a large bill run never returning a bill run summary from the WRLS team. The bill run in question had

- 9,658 transactions
- 2,684 customers

Initially, it was thought to be the sheer number of transactions (we had only tested up to 5K transactions at this point). But after some local testing, we tracked the problem down to the values generated. In this case, the total was more than £22 million.

It's good practise when you need to record money you do so as pence **not** pounds and pence. This nullifies any issues when you come to start performing calculations. The problem for us was the fields in the database had been set as integers. In PostgreSQL the min and max size for an integer is

```
-2147483648 to +2147483647
```

For this particular bill run, we were trying to store a value of `2214000830`. So the generate summary was erroring at the point we were trying to save the result to PostgreSQL.

If we update the database fields to be bigint's then we'll have plenty of room

```
-9223372036854775808 to 9223372036854775807
```

This adds an additional problem though. The underlying Postgres node driver returns these as strings!

So this change covers 2 things

- updating the database fields to be big ints
- ensuring that wherever we extract a billrun record the data is seen by the app as a number and not a string.