Unescape santized html characters (#133)
https://trello.com/c/cjqqnLGs
https://github.com/DEFRA/charging-module-api/pull/119

We added [Disinfect](https://github.com/genediazjr/disinfect) in PR #119 to deal with the fact we were susceptible to XSS attacks.

Disinfect not only deals with malicious content, but it also escapes all content to make it safe for displaying in a browser. This means characters like `&`, `<` and `>` also get escaped. We're not worried about these in our data and feel it's for our clients to determine how to handle them in our responses.

So this change adds a new custom plugin that will unescape payloads after disinfect has escaped them.