Add check for use of `only()` in tests to travis (#134)
The [Hapi Lab](https://github.com/hapijs/lab) test framework comes with the ability to run specific tests by adding `.only` to the describe block. For example

```javascript
describe.only('Sanitizing requests to the API', () => {
  // ...
})
```

This is great! Just don't forget to leave them in when pushing your commits to GitHub. Having not learnt my lesson, this adds a check for use of this to the Travis build.