Add back check for use of `only()` in tests to CI (#151)
https://github.com/DEFRA/charging-module-api/pull/148
https://github.com/DEFRA/charging-module-api/pull/134

In PR #134 we added in a check to our CI for temporary tags. The original description was

---
The [Hapi Lab](https://github.com/hapijs/lab) test framework comes with the ability to run specific tests by adding `.only` to the describe block. For example

```javascript
describe.only('Sanitizing requests to the API', () => {
  // ...
})
```

This is great! Just don't forget to leave them in when pushing your commits to GitHub. Having not learnt my lesson, this adds a check for use of this to the Travis build.
---

When we moved the CI to GitHub actions in PR #148 we lost the check. Having figured how to do it, this adds the check back in.