Remove unused endpoints (#153)
https://trello.com/c/letq5mhW

We had developed some endpoints in the expectation they would be needed by our user. However, after a recent consultation, we have confirmed there are some that are not used or unlikely to be used for a very long time.

So the team have agreed that rather than continue to develop and maintain them we should remove them until they are actually needed.

This will reduce the complexity of the solution, and the maintenance burden on the tech team (development and testing).

This change removes the agreed endpoints and any functionality related to them.

Endpoints removed

> GET /v1/{regimeId}/billruns
> PATCH /v1/{regimeId}/billrun[{billrunId}/unapprove
> POST /v1/wrls/transactions
> DELETE/v1/wrls/transactions/{id}
> GET /v1/wrls/customer_changes
> GET /v1/wrls/customer_changes/{id}

We also had to make changes to the tests. All the test coverage was for adding and removing a transaction. There was nothing for adding and removing a bill run transaction. So we migrated across the existing tests to rectify this. We also covered some scenarios specific to bill run transactions.

As part of this, we ended up creating some new test helpers and introduced the idea of fixtures. There is some additional work to be done to finish this. But that's for another commit!