Refactor billrun & transaction tests helpers (#157)
Whilst working on [Create a PATCH version of billrun approve endpoint](https://github.com/DEFRA/charging-module-api/pull/156) we found we need to refactor the `createBillRun()` to be like the new `createTransaction()` we created recently. We needed the flexibility to set any field rather than the few currently allowed.

When we did the refactoring though it became clear the methods were now almost identical. We were already aware of the `findBillRun()` and `findTransaction()` helpers were also another example of near-identical methods.

So we've decided to take a step back and add this refactoring. Here we add a new database helper. The intention is to bring the more low-level methods into it and use arguments passed in from the other helpers to remove the duplication yet still offer the same functionality.