Create a PATCH version of billrun send endpoint (#156)
The convention in the API is for any endpoint which is used to tell us to do something, for example, 'approve' a bill run, to expect a `PATCH` request. That is apart from 'send' a bill run. That's a `POST`.

Like nails on a blackboard, it causes us pain whenever we see it. So this change is the first step in rectifying the issue. As agreed with the WRLS team this change will add a new endpoint that accepts a `PATCH` request but runs exactly the same code.

We can then ship this without breaking their current system. They will then have time to migrate to using the `PATCH` endpoint. At which point we can delete the existing `POST` version.

No more nails on blackboards 😁