Underscores in urls? Oh my! (#164)
Most likely because that's what the underlying controller/endpoint is called, we seem to have some endpoints that expect an underscore in the path.

It's perfectly permissible. However, it's definitely not the convention for URLs. To quote Google

> Consider using punctuation in your URLs. The URL **http://www.example.com/green-dress.html** is much more useful to us than **http://www.example.com/greendress.html**. We recommend that you use hyphens (-) instead of underscores (_) in your URLs.

So before the API hits a live environment, we're taking the opportunity to follow the current convention of using dashes.

There is also the 'blackboard' effect on us as a delivery team. It's for this same reason we added a [`PATCH` version of the bill run `/send` endpoint](https://github.com/DEFRA/charging-module-api/pull/156). 🧑‍🏫 😁

** Note

Because of the use of the `basePath` variable and `regimeCustomerChangePath()` method, the simplest solution was to simply copy and paste the method and give it a new name.

It does mean we've duplicated 99.99% of the code. But this will only be until the next release version at which point having done it in this way it will be extremely simple and safe to delete.