Fix array and boolean true handling by plugins (#166)
See [Fix payload plugins for arrays and boolean true](https://github.com/DEFRA/sroc-charging-module-api/pull/49)

We found when working on the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) that we have an issue with our plugins that iterate through and make changes to the request payload.

In this project we know the `UnescapePlugin` does not handle arrays correctly. If, for example, the array sent is `["cfd", "wrls"] what we would eventually see is `{"0":"cfd","1":"wrls","2"}`.

We also found that [Disinfect](https://github.com/genediazjr/disinfect) which we added to protect ourselves from XSS attacks cannot handle boolean `true` in a request. It converts them to `'true'` though oddly `false` is left untouched.

In the end, for **sroc-charging-module** we ended up dropping both Disinfect and our individual custom plugins and rewriting a new one. This change brings over a version of that (minus the request cleaning functionality) in order to fix the charging-module-api.