Prevent duplicate transactions (#172)
https://trello.com/c/FkpwxW8V/1866-prevent-duplicate-transactions-better-error-handling

WRLS have requested a change to the add transaction end point to help in situations where adding a transaction fails. When calling the endpoint, they will specify a clientId; if a transaction with this id doesn't already exist then a new transaction will be created, otherwise the transaction id will be returned. This means that a failed transaction can be retried without creating duplicate transactions.

This PR introduces a find_transaction_by_client_id service which can be given a client id and regime, and will return a matching transaction if one exists (or null if it doesn't). This is then used by the add transaction endpoint to check whether it's attempting to create a duplicate transaction and act accordingly.