Fix defect 338 discovered during net zero value testing (#173)
https://trello.com/c/EM0YSGtT/355-338-defect-credit-transaction-in-deminimis-invoice-is-incorrectly-flagged-as-deminimis-false

Deminimis is determined at the invoice level, ie. if an invoice's total value is under £5 (and it is a debit) then deminimis is set to true for this invoice and all its transactions. At some point during zero value transaction work a few months ago, a change was made which had the side-effect of not setting the deminimis flag to true for credits (ie. transactions with a value <0) -- this is incorrect but has only just come to light during testing of the duplicate transactions work. This PR therefore reverts this to how it behaved previously.