Add support for using Docker for local development (#175)
https://github.com/DEFRA/sroc-charging-module-api/pull/60
https://github.com/DEFRA/sroc-charging-module-api/pull/65

Over on the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) we recently added Docker. Specifically, we ensured it was done in such a way that we could develop the app in docker rather than running directly on the host. This was alongside being able to set up an environment quickly and be ready for when we need to deploy the app using Docker.

Though we intend the **sroc-charging-module-api** to eventually replace this, we are still called on to make changes to this project. So keeping the development environments consistent will be a big help to us. Also, as this project is currently set up for deployment to an AWS ECS environment we can use it to work out any wrinkles with the updated Dockerfile we'll be using. We can then migrate what we have learnt back to the **sroc-charging-module-api**.

* Note about the Dockerfile

The one key difference is we are still running the app via npm when it is run in AWS. NPM does not handle SIGTERM gracefully which is why you are advised to run your app via node directly. Or why you will see the use of [tini](https://github.com/krallin/tini) in some containers as a wrapper to ensure SIGTERM is handled gracefully.